// sanity.types.ts
export type SanityImageCrop = {
  _type: "sanity.imageCrop";
  top?: number;
  bottom?: number;
  left?: number;
  right?: number;
};

export type SanityImageHotspot = {
  _type: "sanity.imageHotspot";
  x?: number;
  y?: number;
  height?: number;
  width?: number;
};
// ...
export type PROJECT_BY_SLUG_QUERY_RESULT = {
  _id: string;
  title: string | null;
  slug: string | null;
  year: number | null;
  introduction: string | null;
  information: Array<{
    _key: string;
    label: string | null;
    value: string | null;
  }> | null;
  mainImage: {
    alt: string | null;
    hotspot: SanityImageHotspot | null;
    crop: SanityImageCrop | null;
    asset: {
      _id: string;
    } | null;
  } | null;
  caseContent: Array<
    | {
        children?: Array<{
          marks?: Array<string>;
          text?: string;
          _type: "span";
          _key: string;
        }>;
        style?: "normal";
        listItem?: never;
        markDefs: Array<{
          href: string | null;
          _type: "link";
          _key: string;
        }> | null;
        level?: number;
        _type: "block";
        _key: string;
      }
    | {
        asset: {
          _id: string;
        } | null;
        media?: unknown;
        hotspot: SanityImageHotspot | null;
        crop: SanityImageCrop | null;
        alt: string | null;
        _type: "image";
        _key: string;
        markDefs: null;
      }
  > | null;
} | null;

// Source: sanity/queries/studio.ts
// Variable: STUDIO_QUERY
// Query: *[_id == "studio"][0]{    about[]{      ...,      markDefs[]{ ..., _type == "link" => { href } }    },    clientsCollaborators,    consultants,    contactBlurb[]{      ...,      markDefs[]{ ..., _type == "link" => { href } }    }  }
export type STUDIO_QUERY_RESULT =
  | {
      about: null;
      clientsCollaborators: null;
      consultants: null;
      contactBlurb: null;
    }
  | {
      about: Array<{
        children?: Array<{
          marks?: Array<string>;
          text?: string;
          _type: "span";
          _key: string;
        }>;
        style?: "normal";
        listItem?: never;
        markDefs: Array<{
          href: string | null;
          _type: "link";
          _key: string;
        }> | null;
        level?: number;
        _type: "block";
        _key: string;
      }> | null;
      clientsCollaborators: Array<string> | null;
      consultants: Array<string> | null;
      contactBlurb: Array<{
        children?: Array<{
          marks?: Array<string>;
          text?: string;
          _type: "span";
          _key: string;
        }>;
        style?: "normal";
        listItem?: never;
        markDefs: Array<{
          href: string | null;
          _type: "link";
          _key: string;
        }> | null;
        level?: number;
        _type: "block";
        _key: string;
      }> | null;
    }
  | null;